Add external Ceph cluster support
This patch introduces an API extension to support an external Ceph cluster in
the glance operator.
The GlanceBackend is computed accordingly, and if Ceph is present (which means
that CephMonHosts, CephClusterFSID and CephClienKey have been defined), the rbd
backend is selected, otherwise "file", which is the default, is still used.
The API structure has been extended and now includes the CephBackend parameters
that can be used to set the "rbd" backend when the config file is built.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>