Check if the service is not found during reconcileDelete
This patch adds an error check to the glance-operator during
reconcileDelete: if it's an error and it's not found we can assume the
service has been deleted, otherwise we must fail. The same logic has
been added to the cinder operator, and now we have a well established
pattern to explicitly check and clean the related resources.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>