Makefile: Make deploy work off the bat
If we run `make deploy` we will not get a running Glance operator,
instead we'll get an image pullback error because the default IMG
doesn't work.

To avoid accidentally pushing a development image as the official
openstack-k8s-operators image the `docker-push` rule is not longer
allowed to push to the default image.

This should not be an issue for automation because the GH action to push
to quay.io doesn't use that make target.

This patch modifies it to use a working one and also changes the default
namespace to "openstack".

Since we are now using the same namespace as all the other operators,
we'll remove the namespace creation and deletion from `manager.yaml`.

Also drop `controller:latest` logic from `create_bundle.sh` as this
would no longer get triggered as our default Makefile image is now set
to a fully qualified URL (quay.io....)