Makefile: Install also sets RBAC
The `install` target in `Makefile` only sets the CRDs, which is not
enough when you want to run the operator locally, since the RBACs will
be missing and the operator will not work.

This patch changes the `install` and `uninstall` targets to not only
take care of the CRDs but also of the RBACs.