Cleanup DB sync job after hash is persisted
There was a lib-common issue where job.DoJob() deleted the Job right
after it is succeeded. Then the controller tried to store the hash of
the Job in Status. But if that failed (due to e.g. Conflict) then in the
next Reconcile call lib-common would re-create and re-run the Job as the
hash of the previous Job and the Job itself is lost.

The lib-common is fixed in a way that the automatic Job deletion is
removed from DoJob() and that needs to be now done on the caller side
via the new job.DeleteAllSuccededJobs() call.

This patch adapts to the new lib-common version by doing the Job cleanup
right after the Status.Update() successfully persisted the Job's hash.

Depends-On: openstack-k8s-operators/lib-common#77