Enable keystone authentication for glance
As of now if you deploy glance-operator is doesn't enable authtoken middleware
causing glance to use no-auth, which means glance is not validating any API
calls via keystone. This is also the reason why owner field on image is set
to None if you create glnace image. In order to enable it you need to set
'flavor' as 'keystone' under 'paste_deploy' section of glance-api.conf. Setting
this leads to another issue that under 'keystone_authtoken' section 'services'
project name is used which is not a valid keystone project.

This commit fixes all the issues related to keystone authentication for
glance operator.