Update internal/external service configuration
The show_multiple_locations setting, as well as
show_image_direct_url, needs to be configured for the internal
and external services to work correctly. Both settings are True
for the internal service in order for it to provide image location
data. Conversely, both are set False for the external service,
so that image location data is excluded from public view.