Optimize the Propagation struct
As done for cinder operator, this patch optimize the code to and fix
both ServiceTypes and Groups as consts in the glance code. By doing this
we can avoid redefining the same variable many times in the same code
and we can just call what's defined in a single place.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>