Bump lib-common to get fix in common/job
Glance was using a lib-common version that already implemented the
automatic but delayed job cleanup logic from October, but there was a
bug in the initial implementation of that code that was fixed a week
later in lib-common. As glance was not using that fix when the db sync
Job is automatically deleted after 10 minutes glance started throwing
errors.

This patch bump to the latest lib-common/common module to pick up the
fix.