Set KeystoneAuthURL according to the instance type
Previous changes split the GlanceAPI between internal and external. This
means that we should connect to the Keystone endpoint according to the
nature of the GlanceAPI instance.
This change fixes the way glance-api.conf resolves the endpoint, keeping
it generic as it's injected by the operator when the pods are created.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>