Trim GlanceAPI CRD and optimize extraMounts usage
The goal of this patch is to give an alternative approach to PR#184
where we tried to trim the GlanceAPI CRD as much as possible by
injecting extraMounts from the top-level Glance CR and getting notified
on a change coming from upper layer. However, the approach described in
PR#184 "breaks" the top-down pattern where parameters are specified in
the Spec and passed to the deployment, and in a way it "hides" the data
from the requests. The current patch represents an attempt to solve the
same problem with a different strategy: it trims less than PR#184 but
keeps the pattern unchanged. In particular the GlanceAPISpec is built
within the Glance controller, but it's not explicitly included in the
GlanceAPI. A GlanceAPITemplate, instead, is included and used to define
the top-level input parameters provided and set to the sub-level
resources. By doing this we can still keep the ExtraMounts definition in
the GlanceAPISpec, but we don't replicate the same parameter in the
top-level CRD (as the Spec are not explicitly included).

Signed-off-by: Francesco Pantano <fpantano@redhat.com>