Optimize the way GlanceAPISpec is built
As done for Cinder, this patch reduces the need to repeat the same
fields many times in different structures, especially in the
GlanceAPISpec. In particular, GlanceAPISpec includes a GlanceAPITemplate
which represents a list of parameters defined for the specific instance,
while other parameters (e.g. ExtraMounts or PasswordSelectors) are
inherited from the top level CR.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>