Propagate secretConf to the GlanceAPI pods
The previous patch removed the crudini dependency and introduced a new
parameter, 'CustomServiceConfigSecrets', that can be used to define a
list of secrets (by name) that should be processed by the operator and
the init.sh script and propagated to the glance service.
The previous patch neglected to add the pkg/glance related change that
is supposed to process/mount the resulting volumes, and the purpose
here is to fix that scenario.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>