Rename apiEndpoint key for consistency
Currently some operators use apiEndpoints(plural) while the other use
apiEndpoint(singular). This is quite confusing and requires users to
check which key is presented by a operator, to obtain the same
information.

This renames the key to the plural, because this status field is a map
which can contain multiple endpoints.