Set PVC access mode ReadWriteOnce
This changes the PVC access mode from ReadWriteMany
to ReadWriteOnce to allow the default lvmtopo driver
installed in SNO to provide stoage to glance.

When using the local file backend for glance
each pod uses its own pvc and there is no assumtion that
the glance instnace can share data between the pods via
the filesystem. Infact glance assumes that the backing
store used by the file backend is not shared.