Get System wide admin OSClient when registering quota limits
When the "quotas" struct is passed to the main Glance CR, the operator
needs a system wide scoped token to make sure the request is run
successfully. This change allows the glance-operator to not rely on the
keystone operator to get the right OSClient, and build it based on the
new AuthOps "System" parameter recently added in lib-common.
For this reason this patch also bumps the 'lib-common/modules/openstack'
version to the latest released tag.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>