Introducing Cinder backed for glance
This patch introduces the missing code in the glance-operator to choose
cinder as a backend. As done for TripleO, when cinder is selected as a
backend, we should run GlanceAPI in privileged mode, hence the new bool
has been added to the basic deployment. If cinder is not mentioned in
the "enabled_backends" list, we should still run GlanceAPI as not
privileged.

Co-authored-by: Francesco Pantano fpantano@redhat.com