Run make gowork as part of the postUpgrade tasks
precommit-check seems to fail on multiple renovate related PRs due to
the fact that 'make gowork' is missing. By adding this target we can
make sure to update both modules in the root of the project and in the
api/.
Adding that target also bumps the api/ requirements as part of the
regular precommit-check run.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>