Check if Cinder CR exists when is set as a Glance backend
When Cinder is set as a Glance backend, we might want to wait for Cinder
to appear in the ctlplane before moving with the GlanceAPI deployment.
This change watches for the expected Cinder CR, and it unblocks the loop
as long as it's present in the ctlplane. Note that we're not going to
wait for CinderVolume to be ready, otherwise we risk to block Glance
deployment for an undefined amount of time, compromising the other
backends to start serving requests.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>