Cinder backend: Fix iscsi and multipath
OpenShift deployments are different from TripleO deployments.  In
TripleO we run iscsid and multipathd daemons in containers and their
client side commands are in sync (version-wise) in the service
containers.  It is also required that there are no other daemons running
in the host or other containers.

In the OpenShift world this is different, iscsid and multipathd run on
the host and therefore their version won't be in sync with the one in
the openstack service containers, which means that they won't work
correctly.

The reason why OpenShift runs iscsid and multipathd in the host is so
that they can be shared between the different things that may need it:
OpenShift itself, CSI plugins, and in this case also OpenStack services.

The way to use the host iscsi and multipath services is by exiting the
pod namespace and running the commands on the host namespace using
nsenter.  Even if this may look nasty/weird, this is the recommended way
by the storage OpenShift team and what CSI plugins do.  For this we need
to share the host PID namespace, but just like running Glance container
as privileged we only do it when we know we are using Cinder.

This patch changes how glance operator deploys glance-api to fix the
iSCSI and Multipath support when using Cinder as a backend.

The fix consists in replacing the iscsiadm, multipathd, multipath, and
scsi_id commands with a simple script (`run-on-host`) that calls
nsenter.  The replacements of the binary is done using the kolla config
file.

This allows us to use the same containers in OpenShift and in baremetal
where we run the daemons in a container.

The `run-on-host` script supports 2 ways of replacing commands: being
copied or symlinked. In our case kolla uses the copying mechanism.

Having the binaries replaced when we are not using cinder as a backend,
and therefore are not using the host PID namespace, is not a problem
since Glance is never going to call those binaries.