Cinder backend: Support NVMe-oF transport protocol
This patch adds support for the NVMe-oF protocols by sharing the
/etc/nvme path from the host so that os-brick can generate the nqn and
use it consistenly in all the containers.

It also uses the hosts' `nvme` command (using `nsenter`) instead of the
one in the cinder containers.  This may not be a hard requirement right
now, as command in the containers could be compatible, but this could
change in the future since these are 2 different Operating Systems:
CoreOS and UBI/RHEL.