Do not ship policy.yaml by default
policy.yaml atm is stored in a secret, it can't be changed by any human
operator because glance-operator will restore it and it grows a lot the
size of the GlanceAPI secret.
This patch removes it from the Secret, kolla won't sync it by default
but it can still be provided via extraVolumes, injecting the oslo_policy
section via customConfigService override.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>