Create -scripts as top-level resource
GlanceAPIs have no reasons to regenerate the -scripts mounted to the
Pods. The scripts represent a resource that can be inherited from the
top level CR.
This patch let the glance_controller generate this resource: it will
be mounted (if Cinder is used as a backend) by all the GlanceAPI pods.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>