Create an intermediate struct to tie together internal and external API
GlanceAPI is made by two main components: InternalAPI and ExternalAPI.
Right now the two deployments are processed independently by the main
Glance controller, and in terms of exposed API they represents two
different instances logically coupled together.
This patch introduces a change in the main Glance API to provide an
intermediate struct that appears in the main GlanceSpec and defines
the two internal and external instances inside. This is the first
step to move to a more flexible structure that allows the deployment
of an arbitrary number of GlanceAPIs, or just deploy a single instance
if file is used as a backend.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>