Allow to deploy a single GlanceAPI instance
Today the GlanceAPI instances are hardcoded in the basic GlanceSpec
template.
This strategy works well because we can always ensure a distinction
between Internal and External, but it's not flexible and any change
in the underlying GlanceAPI Sub-CRs requires the main API to change
in its structure.
The solution proposed by this patch makes the split between internal
and external instances implicit, and it maintains at the same time
backward compatibility with the logic associated to the previous
implementation.
A new 'Type' field has been introduced within the GlanceAPIInstance
struct: the purpose of this field is to associate the right behavior
to the deployed API, and it characterizes the deployment in that
sense.
GlanceAPI are still deployed in "split" mode by default, unless the
spec specifies something different.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>