Do not pass plain string to cronjobDef
This patch creates an enum that can be used to reference the cronJobType
during the Job creation.
By doing thig we are not allowed to pass plain strings to the CronJob
function and mess up with the object without control.
In addition, the cronJob schedule should be in line with the Job type
and assigned before creating the CronJob object.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>