Improve envTest including PVC asserts
It's a good idea assert that a local pvc is created and exists when the
main Glance CR is applied. This patch adds the related assert functions
and check that a local PVC exists when Glance is created. Given that we
are not enabling image-cache in the 'DefaultGlanceSpec', an additional
assert that checks that a '-cache' pvc has not been provisioned during
the creation phase.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>