Replace DBPurge Debug boolean with CronJob
We previously introduced a DBPurge boolean under the GlanceDebug struct
to make sure we're able to trigger a cronJob in debug mode. Since the
first patch multiple cronjobs have been added (pruner and cleaner),
hence it makes sense to change the boolean variable name to something
that is valid for a generic cronjob object.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>