Do not use MustParse to process PVC size
The resource.MustParse function was previously used by the operator to
process the storageSize request provided by the top-level Glance CR.
However, even though defer() and recover() were used to handle the
panic() generated by that function in case of a wrong user input, that
approach doesn't seem to work properly and the operator crashes without
any chance to recover.
The MustParse implementation wraps the ParseQuantity function, that
instead of panic() returns an error that can be easily catched at the
operator level.
This change moves away from the MustParse usage and rely on the
ParseQuantity mechanism.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>