Allow to deploy a list of glanceAPIs
This patch is the first of a series where the idea is to refine the
mechanism of deploying a list of GlanceAPI.

The goal is to cover two main use cases:
- DCN, where each glanceAPI instance can be tied to a different backend
  (e.g., multiple Ceph clusters in case of Storage at Edge)
- Dedicated workload: it might be possible to reserve instances for a
  given workload, and from a deployment perspective, multiple glanceAPI
  instances is the first step to achieve that goal

The functional tests have been updated to be compatible with a list of
glanceAPI, and kuttl will be updated as a follow up patch.
Finally, more logic has been moved from the existing controllers to the
commmon area.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>