Add an annotation to select which API should be registered
Now that multiple glanceAPI can be deployed, we need a mechanism to
select which one should register its keystone endpoints. The umbrella
glance controller should be responsible to drive it via the main CR;
for this reason a new field has been introduced to select (by name)
which one should have its endpoints registered in the catalog.
When multiple glanceAPI are deployed, the "ensureKeystoneEndpoints"
function takes care about setting the right condition, so we can mark
as valid the use case where the endpoints are not registered.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>