Add ensureDeletedEndpoints as a separated function
As a result of moving the KeystoneEndpoint cleanup in a separated
function, this change actually solves a couple of critical issues
associated to the way keystone endpoints are exposed.
In general, for Glance, there's an "image" service present in the
keystone catalog, and it has an associated set of endpoints that
are used as the user entrypoint when the interaction with the
glance cli takes place (tipically public and internal endpoints).
There is a 1:1 relation between the endpoints and the image service:
we can't have more than one image service, as well as we can't have
more than one set of endpoints associated to the image service.
For this reason the logic of exposing the endpoints now takes into
account an "apiAnnotation" that acts as a selector, and it deletes
any pre-existing "keystoneendpoint" CR before the new set of
endpoints is rolled out, updating the keystone catalog.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>