Replace keystoneBackend variable and update envTests
This patch updates the usage of keystoneEndpoint variable and fixes the
envTest to make sure the glanceAPI that is supposed to be registered to
keystone has the associated annotation.
The main Glance webhook has been improved to catch situations where the
keystoneEndpoint has no associated instances. In that case an error is
raised at creation time, and the human operator can fix the CR.
In addition, given that we're increasing the ways glance can be deployed,
a "samples/layout" directory has been created, and the samples have been
updated and can be applied using kustomize as a general approach.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>