Update kuttl tests to work with the new API definition
Kuttl tests have been updated to reflect the new GlanceAPI definition.
As an additional point, the copy && deploy pattern has been removed and
now, for each sample layout in config/samples an associated kuttl test
can be built and run.
So far only single and multiple have been moved, but the idea is to have
a follow up change to build edge, base and multiple.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>