Rework the logic that handles the apiDeployment
The logic that handles the apiDeployment has been reworked and moved to
a dedicated function.
By doing this we're able to clearly have evidence in the code of the
flow in the main reconcile loop.
In addition, webhooks are updated and do not prevent the deployment of
additional glanceAPI later in time.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>