Handle glanceAPI cleanup
We have the ability to grow the deployment and add more glanceAPI
instance as needed, as well as select which one should be used as
the main instance registered to keystone catalog. With this patch
we have now the ability to cleanup an instance thar no longer
belongs to the main CR. The endpoints registered in the Status are
removed as well.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>