Fix labels and selectors in the apiDeployment flow
This patch fixes a few issues in the label and selectors assignment and
goes back to the previous deployment workflow so we don't have too much
wrapping in the existing logic.
The delete flow has been updated as well: the Finalizer that prevented
the glanceAPI to be removed is gone and the glanceAPICleanup is not run
within the for loop.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>