Update extraVolumes to match the instance propagation
Now that we have the ability to define multiple Glance instances, we
can propagate extraVolumes not only to the Glance components (or,
logically, the GlanceAPI subgroup), but we can do the same thing by
specifying a defined instance (e.g. api0). As done for both Cinder
and Manila (where this mechanism is already in place), the code has
been updated to reflect this way of propagating volumes.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>