Detect invalid backend configuration
When the GlanceAPI is split between internal and external instances, if File is
used as a backend Glance won't work as expected. Each instance is going to have
its own RWO pvc where the images are stored, and it might happen that while an
image is uploaded via the external instance, an application might request it via
the internal instance (which is associated to the keystone internal endpoint),
resulting in a failure because "no data are associated to the requested image".
For the reason above, when the "File" backend is used, the operator should
prevent the API split (type: single), or block any attempt to use the File
backend with this layout.
This patch introduces a webhook related improvement that validates the layout
and fails if some criteria are not met.
In particular, if a split (internal + external instances) layout is used
with a File backend, the ValidateCreate() function raises an exception
preventing the deploy.
Other than that, we know that if an empty spec is applied, a "default" GlanceAPI
instance is created. The Default() function now enforces a "split" layout, it
sets "replicas: 0" and creates a "CHANGE_ME" dummy backend.
This will force the human operator to think in advance about the backend that
should be associated to Glance as we don't want File to be the default.
Both the samples and the functional tests are aligned with the new webhook funcs
to make sure we don't get a "Detected invalid backend" issue.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>