Update the NFS backend example
This patch introduces the missing documentation for a NFS backend.
It includes both an updated CR that is consistent with the extraMounts
usage in the OpenStackControlPlane, and it introduces a script to help
testing this backend with crc.
There are also instructions and an example that shows how to double check
the NFS backend is actually used by the GlanceAPI Pod.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>