Fail during Validate Create and Update if replicas > 0
It is possible that the human operator has not chosen the backend yet.
With this update we can still move forward with the Glance deployment
and not block the GlanceAPI creation if replicas == 0.
This means that if File is detected as a backend (because no other
choice has been made) and the layout type is set to "split", fail only
if the glanceAPI defined in the CR has replicas > 0, otherwise move on
and perform both database creation and dbsync, as well as the API
creation.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>