Fix image cache related cronJobs
This patch fixes the way cronJobs are created. With a particul focus to
the image-cache feature, the goal is to perform a cronJob creation
passing the right "glance-image" pvc resulting from the current
statefulset creation.
Each GlanceAPI instance (either a replica or a dedicated API) needs an
associated cronJob when the ImageCache is defined. A dedicated cronJob
is needed because the glance-{cleaner,pruner} tools act against a path
(/var/lib/glance/image-cache) mounted for a particular cache pvc created
by the StatefulSet as a VolumeClaimTemplate.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>