Expose DBPurge Age and Schedule
This patch updates the current glance CR to expose both dbPurgeAge
and dbPurgeSchedule as customizable parameters. They were previously
hardcoded as constants in the code, and it makes sense removing that
part as long as the human operator has the ability to customize them.
This also aligns the glance-operator with the work already done in
both Manila and Cinder.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>