Add distributed_image_import support
This patch introduces the first items required by the glance-operator to
support distributed_image_import. Distributed image import requires > 1
glanceAPI replicas and it is possible to redirect the import operator
to a given worker via worker_self_reference_url. However, this means
that we need to resolve the current POD Ip address. StatefulSet provides
stable network access to the Pod/container/replicas, and by adding a
headless service we can resolve the Pod even though it's scheduled on
a different node.
This patch, however, is not sufficient to have a fully working feature,
because we still need some logic within the operator to inject the right
worker_self_reference_url to a given replica that reflects the current
fully qualified hostname.
By adding the "Domain" parameter, we keep track of the cluster domain
resolved by kube-dns that is the base to resolve the POD via hostname.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>