Configure per Pod/replica set_worker_self_url option
This patch fixes the glanceapi_controller implementation to propagate
and support the worker_self_reference_url config option. It relies on
kolla_extend_start script (that is currently empty within the pod) to
get the current hostname and set it as global option (01-config.conf),
inject to the Glance service.

[1] https://github.com/openstack-k8s-operators/docs/blob/main/service_config.md#note-1

Signed-off-by: Francesco Pantano <fpantano@redhat.com>