Create an headless service associated to each glanceAPI statefulset
We can't turn the existing internal/external Services into headless
services, because they are used to expose the component and we can
pass overrides for metallb and HA strategies (e.g. type: loadbalancer)
when network isolation is used. For this reason, this patch creates
an additional headless service for each statefulset we expose. By
doing this we're able to keep the existing model and, at the same time,
resolve pods by hostname.
The "glance-direct" import method is enabled by default, hence we do
not require any change in the glanceAPI config.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>