Adjust functional tests to reflext the headless service change
In order to be able to resolve pods by hostname, we need to deploy
an headless service associated to each StatefulSet. The headless
service name must match with the StatefulSet Name and ServiceName,
hence this patch improves the functional tests to make sure this
behavior is reflected in the current implementation.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>