Fix Ceph image conversion
With the current implementation, we require the human operator to build
a PVC for Image Conversion purposes. The PVC, attached to an arbitrary
storage class, can then be injected and mounted to the right path via
ExtraMounts. However, this implementation is good only for RWX PVCs,
because it realizes a common space for all the glanceAPIs and replicas
where the conversion happens. If a RWO PVC is used, a replica deployed
on a different node won't be able to mount the PVC, and worse, it will
be stuck in `Pending` state due to the Distribution affinity policy
implemented in the operator. Even with the NodeSelector usage, there
will be cases where this is going to generate issues.
This patch fixes the problem described above with the following:

1. It introduces a PVC, created by the StatefulSet (via templating)
   only for the "external" glanceAPI instances
2. It creates the imageConversion PVC only if Ceph is enabled
3. It automatically sets the Image conversion plugin feature in Glance
   00-config.conf when Ceph is part of the enabled_backend string
4. Takes care about the StatefulSet update for fields that are forbidden

The image conversion PVC is a RWO, stable storage that can be used by
each replica (tested with replica:3) and we avoid any manual action
required by a human operator. ImageConversion is enabled by default with
Ceph and it's transparent to the end user.

Fixes: https://issues.redhat.com/browse/OSPRH-5690

Signed-off-by: Francesco Pantano <fpantano@redhat.com>