Add missing conditions in both the top-level and glanceAPI controller
Most of the code ofter returns ctrl.Result{}, err in case there's
something wrong with a specific component that is supposed to be
ready before move forward with the reconcile loop. However, an
aspect of this evaluation is the missing update of the condition
for that particular component. This commit fixes it and introduces
the missing updates during the reconcile loop.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>