Fix updates to a condition's LastTransitionTime
Do not update a condition's LastTransitionTime unless its state actually
changes. This ensures the LastTransitionTime isn't always modified
during every reconcilation.
In general, when a reconciliation starts, it keeps track of the previous
conditions, and at the end of the loop the conditions array is updated.
However, if a condition didn't change during the last reconciliation
(condition.HasSameState() from lib-common helps on that front), we
shouldn't update the LastTransitionTime, and we can keep a consistent
view of the conditions at each reconcile.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>