Update the ReadyCondition for GlanceAPI and mirror to the top level CR
This patch updates the Ready condition by removing it from the defer
function and put it at the end of the reconcile loop. In addition,
before marking the GlanceAPI Deployment Ready, it checks that the
number of replicas is equals to the ReadyCount. If the deployment is
still in progress, of this condition is not met, it's mirrored to
the top-level CR that shows at which point the Deployment is blocked.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>