Predict the GlanceAPI volumes order
When TLS is enabled at Pod level (which is the new default introduced by
the openstack-operator), and a statefulset is created, a new revision is
rolled out because of the overrides passed by the OpenStack operator to
the service CR. In glance this introduced an additional issue: in case
of multiple APIs, an iteration is performed through the Spec instances,
and the TLS override is checked out for each endpoint. No one ensures
that the StatefulSet has the same order of the provided mountpoints, and
this might generate multiple (random) rollouts until it converges with
two subsequent revisions that keep the same order. To avoid multiple
restarts, this patch sorts the iteration on the resulting endpoints
associated with the GlanceAPI StatefulSet. By doing this we can always
predict the mount order of Volumes and VolumeMounts and avoid unnecessary
restarts.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>