Move both -httpd and -api containers to dumb-init
As noticed in other operators and for the -log sidecar container, SIGTERM
is ignored on delete. This patch moves the glance-operator to entirely
use dumb-init to make sure the SIGTERM is handled properly when it's
received. In addition, this helps to quicly recreate the API containers
when tls-e is enabled at Pod level.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>