Improve how the memcached instance is handled
Instead of flooding the log with multiple stacktraces, we can gracefully
handle the case where the memcached instance is not ready, log a message
(other than setting the Condition) and requeue the request. If something
changes after the first bootstrap, we should trigger a reconcile loop
in the underlying controller that is supposed to deploy the API using
that memcached instance. For this reason a similar check is added within
the glanceapi_controller, that now watches for the same resource.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>