Simplify PVC interface
Do not automatically create an image-conversion PVC when Ceph is set as
a backend. It will be documented that the requestSize should be bigger
for the default PVC or other strategies should be used (e.g. NFS
share attached to the Pods via extraMounts). This patch removes the
automation that automatically created an image-conversion PVC.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>