Simplify the Storage interface for GlanceAPI
We currently provide both StorageClass and StorageRequest that can be
propagated to the underlying instances.
However, it might be possible that operators use extraMounts to plug a
NFS share mapped to /var/lib/glance.
In this case, the API should be updated to not provision any PVC and
let the human operator to take care about it using the extraMounts
interface.
This patch adds a "Storage" struct that is supposed to act as the high
level interface where the storage strategy for glance is specified.
If the ephemeral boolean is set, no PVCs are provisioned.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>