Fix kuttl TestAsserts for glanceAPI endpoints
We currently have issues in the GlanceAPI TestAsserts because they do not
take into account the api name. As done for tls in the past, this patch
changes the go-template to match the pattern <apiName>-<endpointType>.
However, given the "-" character is not allowed, a workaround is used
with the index on the .status.apiEdnpoints.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>