Annotate the glance SVCs with the associated headless service
When a Glance SVC is created, an associated headless service is
generated: this is used to make glance-direct work and enable
distributed image import.
However, when certificates are created, we need to include the
fqdn related to the headless service, otherwise instances of the
same Pod (replicas) are able to resolve each other, but not able
to proxy requests via HTTPS.
This change annotates the created SVC with the corresponding
headless service, and the openstack-operator can look at this
annotation to build a list of additional SubjectNames that
needs to be included in the CertificateRequest.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>